% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflamingo.data_prepare.R
\name{tflamingo.data_prepare}
\alias{tflamingo.data_prepare}
\title{tflamingo.data_prepare}
\usage{
tflamingo.data_prepare(
  code_path,
  input_folder,
  low_res,
  high_res,
  outputs_folder,
  chr_name,
  assembly
)
}
\arguments{
\item{code_path}{code path of tFlamingorLite}

\item{input_folder}{Path to the scHiC data folder.}

\item{low_res}{low resolutuion backbone}

\item{high_res}{high resolution}

\item{outputs_folder}{Path to the folder to store intermediate outputs.}

\item{chr_name}{Name of the chromosome}

\item{assembly}{Genome assembly version}
}
\description{
Prepare data for tensor completion
}
