% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflamingo.main_func.R
\name{tflamingo.main_func}
\alias{tflamingo.main_func}
\title{tflamingo.main_func}
\usage{
tflamingo.main_func(
  outputs_folder,
  idx,
  input_PD_high,
  input_IF_high,
  input_PD_low,
  input_IF_low,
  domain_res,
  frag_res,
  chr_name,
  nThread,
  sample_rate,
  lambda,
  max_dist,
  error_threshold,
  max_iter
)
}
\arguments{
\item{outputs_folder}{Folder to store intermediate outputs}

\item{idx}{Cell index of the cell currently being calculated}

\item{input_PD_high}{high-resolution input pairwise distance matrix}

\item{input_IF_high}{high-resolution input interactive frequency matrix}

\item{input_PD_low}{low-resolution input pairwise distance matrix}

\item{input_IF_low}{low-resolution input interactive frequency matrix}

\item{domain_res}{Size of the domains in bps, e.g. 1e6. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{frag_res}{Size of the fragment in bps, e.g. 1e4. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{chr_name}{Name of the chromosome, e.g. chr1. Try strawr::readHicChroms() to see available chromosomes.}

\item{nThread}{Number of thread avalable for the reconstruction. Default = 1.}

\item{sample_rate}{Fraction of available entries in Hi-C to be used during the reconstruction. Default = 0.75.}

\item{lambda}{Weights for all sampled entries. Default = 10.}

\item{max_dist}{Maximum allowed distance betwee two consecutive points. Default = 0.01}

\item{error_threshold}{Error thresholds for reconstruction. Default = 1e-3.}

\item{max_iter}{Maximum iterations. Default = 500.}
}
\value{
A data.frame containing the FLAMINGO predicted 3D structure.
}
\description{
Main function to run FLAMINGO part in Tensor-FLAMINGO
}
\keyword{flamingo_main}
